tpmfp: Logarithmic binning works, but not at the low end?
gadget:

Added a sanity check on the truncated shortrange force.  Used this to make educated
guesses for constants required in changing units.  There are big problems in the way I've
distributed the constants across the various factors: namely that theres a term in the
suppression where the limit goes to zero, but on term diverges exponentially.  The
computer chokes of course.  Going to try to redistribute the units